Post comment 2.1 Fixed static set userid

+++ Vextech API/Models/ProductReviewModel.cs
@@ -5,7 +5,7 @@ namespace Vextech_API.Models
     {
         public ulong ID { get; set; }
         public int ProductID { get; set; }
-        public int UserID { get; set; }
+        public ulong UserID { get; set; }
         public string Comment { get; set; }
         public int Rating { get; set; }
         public DateTime Date { get; set; }
